refactor: 파일, 디렉토리 구조 수정, 분리
<before>
- Car.js
- contants.js
- templete.js
- utils.js

<after>
- configs/
  - constants.js
  - dom.js
- models/
  - Car.js
- templete/
  - templete.js
- utils/
  - utils.js
