+++ src/js/configs/dom.js
@@ -1,0 +1,34 @@
+export const SELECTOR = {
+  $APP: '#app',
+  $CAR_NAME_INPUT: '#car-name-input',
+  $CAR_NAME_BUTTON: '#car-name-button',
+  $RACING_COUNT_INPUT: '#racing-count-input',
+  $RACING_COUNT_BUTTON: '#racing-count-button',
+  $RACING_RESULT: '#racing-result',
+  $RESULT_LIST: '#result-list',
+  $RACING_CAR: '.racing-car',
+  $CAR_NAME: '.car-name',
+  $PROGRESS_LIST: '.progress-list',
+  $PROGRESS: '.progress,',
+  $RESULT: '#result',
+  $RESULT_MESSAGE: '#result-message',
+  $WINNERS: '#winners',
+  $RESTART_BUTTON: '#restart-button',
+};
+
+export const DOM_STRING = {
+  CAR_NAME_INPUT: 'car-name-input',
+  CAR_NAME_BUTTON: 'car-name-button',
+  RACING_COUNT_INPUT: 'racing-count-input',
+  RACING_COUNT_BUTTON: 'racing-count-button',
+  RACING_RESULT: 'racing-result',
+  RESULT_LIST: 'result-list',
+  RACING_CAR: 'racing-car',
+  CAR_NAME: 'car-name',
+  PROGRESS_LIST: 'progress-list',
+  PROGRESS: 'progress',
+  RESULT: 'result',
+  RESULT_MESSAGE: 'result-message',
+  WINNERS: 'winners',
+  RESTART_BUTTON: 'restart-button',
+};
